+++ CHANGES.rst
@@ -34,6 +34,9 @@ Features added
 Bugs fixed
 ----------
 
+* #14342: intersphinx: Fix leaking basic auth credentials in error and
+  redirect log messages.
+  Patch by Joshua Swanson
 * #14189: autodoc: Fix duplicate ``:no-index-entry:`` for modules.
   Patch by Adam Turner
 * #13713: Fix compatibility with MyST-Parser.